comment

+++ src/user/user.service.ts
@@ -5,35 +5,27 @@ import { CreateUser } from '../user/dto/user.create'
 import { UpdateUser } from '../user/dto/user.update'
 import { REQUEST } from '@nestjs/core';
 import { Request } from 'express';
+import { LoginCredential } from 'src/auth/dto/login-credential.dto';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class UserService {
   constructor(
+    // @Inject(REQUEST) private readonly request: Request,
+    @Inject(REQUEST) private request: Request,
     private readonly userRepository: UserRepository,
   ) { }
-
-  // 모든 회원
-  getAllUser(): Promise<UserEntity[]> {
-    return this.userRepository
-      .createQueryBuilder()
-      .getMany();
-  }
 
   findOne(user_id: string): Promise<UserEntity> {
     return this.userRepository.findOne({ where: { user_id } });
   }
 
-  //회원조회
-  async getUser(mem_id: number) {
-    // const user =  await this.userRepository.findOne(user_id);
-    // const mem_id = this.request['user']['id'];
-    // console.log(mem_id);
-    const data = await this.userRepository
-      .createQueryBuilder('user')
-      .leftJoinAndSelect('user.contents', 'contents')
-      .where(`user.mem_id = ${mem_id}`)
-      .getOne();
-    return data;
+  // 모든 회원
+  getAllUser(): Promise<UserEntity[]> {
+    console.log(this.request);
+    return this.userRepository
+      .createQueryBuilder()
+      .getMany();
   }
 
   //회원 가입
@@ -43,6 +35,27 @@ export class UserService {
     return user;
   }
 
+  //회원 게시글 조회
+  async getUser_Content(mem_id: number) {
+    const data = await this.userRepository
+      .createQueryBuilder('user')
+      .leftJoinAndSelect('user.contents', 'contents')
+      .where(`user.mem_id = ${mem_id}`)
+      .select(['user.mem_id', 'user.user_id', 'contents'])
+      .getOne();
+    return data;
+  }
+
+  //댓글 보기
+  async getUser_Comment(mem_id: string) {
+    const comments = await this.userRepository
+      .createQueryBuilder('user')
+      .leftJoinAndSelect('user.comments', 'comments')
+      .where(`user.mem_id = ${mem_id}`)
+      .select(['user.mem_id', 'user.user_id', 'comments.comment_text'])
+      .getMany();
+    return comments
+  }
 
   //회원 정보 수정
   async updateUser(id: number, userData: UpdateUser) {
@@ -56,13 +69,18 @@ export class UserService {
   }
 
   //회원 탈퇴
-  async deleteUser(mem_id: number) {
-    await this.userRepository
-      .createQueryBuilder()
-      .delete()
-      .from(UserEntity)
-      .where("mem_id=:mem_id", { mem_id })
-      .execute();
-    return '탈퇴가 완료되었습니다.'
+  async deleteUser(mem_id: number, req) {
+    const loginUser = req.user.id
+    if (mem_id == loginUser) {
+      await this.userRepository
+        .createQueryBuilder()
+        .delete()
+        .from(UserEntity)
+        .where("mem_id = :mem_id", { mem_id })
+        .execute();
+      return '탈퇴가 완료되었습니다.'
+    } else {
+      return '로그인 정보가 일치하지 않습니다.'
+    }
   }
 }
